requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
PyYAML==6.0.1
selenium==4.18.1
pandas==2.2.0
flask==3.0.0